Implement ShareAxisTransition with horizontal/vertical options
